refactor: Section과 Station Join로직을 분리
